Sample Quadrat Data (0.6m by 0.9m)

A data_frame that can be used with the Simple Cleaning Quadrat Data and the Cleaning and Cropping Quadrat Data vignettes to show how quadrat data can be cleaned to produce easy to analyze data frames. Data was collected by the Baum Lab and Kiritimati Field Teams. Data is the uncleaned version of data found in Maucieri and Baum 2021. Biological Conservation. The data are from photo quadrats (0.9m by 0.6m) which were randomly annotated with 54 random points each. At each of these annotated points, the substrate was identified. Photo quadrats were collected on Kiritimati Island in the Republic of Kiribati and document coral cover over time and space. The annotations and output of the data_frame were produced using CoralNet and all annotations were done manually, by trained researchers.

Usage

softcoral_SQuads

Format

A data_frame with 135 columns, which are:

Image.ID

Photo quadrat image ID from CoralNet.

Image.name

The photo quadrat image name.

Annotation.status

If the quadrat has been completely annotated, or if there are more annotations to occur.

Points

The total number of annotation points in the quadrat.

AcCor

Percent of total annotated points annotated for Acropora (corymbose morphology).

AcDig

Percent of total annotated points annotated for Acropora (digitate morphology).

Acr_arb

Percent of total annotated points annotated for Acropora (arborescent morphology).

Acrop

Percent of total annotated points annotated for Acropora.

AcroTab

Percent of total annotated points annotated for Acropora (tabulate morphology).

Astreo

Percent of total annotated points annotated for Astreopora.

B_Acr_arb

Percent of total annotated points annotated for bleached Acropora (arborescent morphology).

B_Acro

Percent of total annotated points annotated for bleached Acropora.

B_Astre

Percent of total annotated points annotated for bleached Astreopora.

BAT

Percent of total annotated points annotated for bleached Acropora (tabulate morphology).

B_Cosc

Percent of total annotated points annotated for bleached Coscinarea.

B_Echin

Percent of total annotated points annotated for bleached Echinophyllia.

B_FavHal

Percent of total annotated points annotated for bleached Favites halicora.

B_Favia

Percent of total annotated points annotated for bleached_Favia.

B_FaviaM

Percent of total annotated points annotated for bleached Favia matthaii.

B_FaviaS

Percent of total annotated points annotated for bleached Favia speciosa.

B_FaviaSt

Percent of total annotated points annotated for bleached Goniastrea stelligera.

B_Favites

Percent of total annotated points annotated for bleached Favites.

B_FavPent

Percent of total annotated points annotated for bleached Favites pentagona.

B_Fung

Percent of total annotated points annotated for bleached Fungia.

BGard

Percent of total annotated points annotated for bleached Gardineroseris.

B_GonEd

Percent of total annotated points annotated for bleached Goniastrea edwardsi.

B_Herpo

Percent of total annotated points annotated for bleached Herpolitha.

B_HYDNO

Percent of total annotated points annotated for bleached Hydnophora.

B_HyExe

Percent of total annotated points annotated for bleached Hydnophora exesa.

BlAcro-Cor

Percent of total annotated points annotated for bleached Acropora (corymbose morphology).

B_Lepta

Percent of total annotated points annotated for bleached Leptastrea.

B_Lepto

Percent of total annotated points annotated for bleached Leptoseris.

Blisop

Percent of total annotated points annotated for bleached Isopora.

B_Lobo

Percent of total annotated points annotated for bleached Lobophyllia.

BlTurbFol

Percent of total annotated points annotated for bleached Turbinaria (foliose morphology).

B_MOEN

Percent of total annotated points annotated for bleached Montipora (encrusting morphology).

B_MOFO

Percent of total annotated points annotated for bleached Montipora (foliose morphology).

B_Monta

Percent of total annotated points annotated for bleached Montastraea.

B_Monti

Percent of total annotated points annotated for bleached Montipora.

B_Oxyp

Percent of total annotated points annotated for bleached Oxypora.

B_Paly

Percent of total annotated points annotated for bleached Palythoa.

B_PaveDUER

Percent of total annotated points annotated for bleached Pavona duerdeni.

B_Pavona

Percent of total annotated points annotated for bleached Pavona.

B_PEYDO

Percent of total annotated points annotated for bleached Pocillopora eydouxi.

B_Plat

Percent of total annotated points annotated for bleached Platygyra.

B_PMEAN

Percent of total annotated points annotated for bleached Pocillopora meandrina.

B_Pocillo

Percent of total annotated points annotated for bleached Pocillopora.

B_Porit

Percent of total annotated points annotated for bleached Porites.

B_Psam

Percent of total annotated points annotated for bleached Psammocora.

B_PVAR

Percent of total annotated points annotated for bleached Pavona varians.

B_Sando

Percent of total annotated points annotated for bleached Sandolitha.

B_UnkCoral

Percent of total annotated points annotated for bleached unknown hard coral.

Cirr

Percent of total annotated points annotated for Cirrhipathes.

COSC

Percent of total annotated points annotated for Coscinaraea.

ECHIN

Percent of total annotated points annotated for Echinophyllia.

Fav

Percent of total annotated points annotated for Favites.

FavHal

Percent of total annotated points annotated for Favites halicora.

Favites halicora

Percent of total annotated points annotated for Favia.

FaviaM

Percent of total annotated points annotated for Dipsastraea matthaii.

FaviaS

Percent of total annotated points annotated for Favia speciosa.

FaviaSt

Percent of total annotated points annotated for Favia stelligera.

FavPent

Percent of total annotated points annotated for Favites pentagona.

Fung

Percent of total annotated points annotated for Fungia.

Gardin

Percent of total annotated points annotated for Gardineroseris.

GonEd

Percent of total annotated points annotated for Goniastrea edwardsi.

Herpo

Percent of total annotated points annotated for Herpolitha.

HYDNO

Percent of total annotated points annotated for Hydnophora.

HyExe

Percent of total annotated points annotated for Hydnophora exesa.

Isopora

Percent of total annotated points annotated for Isopora.

Lepta

Percent of total annotated points annotated for Leptastrea.

Lepto

Percent of total annotated points annotated for Leptoseris.

Lobo

Percent of total annotated points annotated for Lobophyllia.

*MOEN

Percent of total annotated points annotated for Montipora (encrusting morphology).

*MOFO

Percent of total annotated points annotated for Montipora (foliose morphology).

Monta

Percent of total annotated points annotated for Montastraea.

Monti

Percent of total annotated points annotated for Montipora.

Oxyp

Percent of total annotated points annotated for Oxypora.

Paly

Percent of total annotated points annotated for Palythoa.

PaveDUER

Percent of total annotated points annotated for Pavona duerdeni.

Pavon

Percent of total annotated points annotated for Pavona.

PEYDO

Percent of total annotated points annotated for Pocillopora eydouxi.

Plat

Percent of total annotated points annotated for Platygyra.

Plero

Percent of total annotated points annotated for Plerogyra.

PMEAN

Percent of total annotated points annotated for Pocillopora meandrina.

Pocill

Percent of total annotated points annotated for Pocillopora.

Porit

Percent of total annotated points annotated for Porites.

Psam

Percent of total annotated points annotated for Psammocora.

PVAR

Percent of total annotated points annotated for Pavona varians.

Sando

Percent of total annotated points annotated for Sandolitha.

Tuba

Percent of total annotated points annotated for Tubastrea.

TURB

Percent of total annotated points annotated for Turbinaria.

UnkCoral

Percent of total annotated points annotated for unknown hard coral.

ANEM

Percent of total annotated points annotated for sea anemone.

B_Clad

Percent of total annotated points annotated for bleached Cladiella.

B_Sinu

Percent of total annotated points annotated for bleached Sinularia.

Clad

Percent of total annotated points annotated for Cladiella.

EncBry

Percent of total annotated points annotated for encrusting Bryozoan.

EUR

Percent of total annotated points annotated for sea urchin.

HYDCO

Percent of total annotated points annotated for Hydrocoral.

Hydra

Percent of total annotated points annotated for Hydroid.

Mille

Percent of total annotated points annotated for Millepora.

MOBI

Percent of total annotated points annotated for bivalves.

Sarco

Percent of total annotated points annotated for Sarcophyton.

SECO

Percent of total annotated points annotated for sea cucumber.

Sinu

Percent of total annotated points annotated for Sinularia.

Sponge

Percent of total annotated points annotated for sponge.

Stylas

Percent of total annotated points annotated for Stylaster.

UnkTUN

Percent of total annotated points annotated for unknown Tunicate.

XmasW

Percent of total annotated points annotated for Christmas Tree Worm.

ZOAN

Percent of total annotated points annotated for Zoanthid.

B_Sarco

Percent of total annotated points annotated for bleached Sarcophyton.

Sand

Percent of total annotated points annotated for sand.

Sediment

Percent of total annotated points annotated for sediment.

SCRO

Percent of total annotated points annotated for consolidated rock.

B_Loph

Percent of total annotated points annotated for bleached Lobophytum.

CYAN

Percent of total annotated points annotated for Cyanobacteria films.

Loph

Percent of total annotated points annotated for Lobophytum.

Rubble

Percent of total annotated points annotated for broken coral rubble.

SHAD

Percent of total annotated points annotated for shadow.

Trans

Percent of total annotated points annotated for transect hardware.

Unc

Percent of total annotated points annotated for unclear.

AVRA

Percent of total annotated points annotated for Avrainvillea.

Caul

Percent of total annotated points annotated for Caulerpa.

CCA

Percent of total annotated points annotated for crustose coraline algae.

Dict

Percent of total annotated points annotated for Dictyota.

DICTY

Percent of total annotated points annotated for Dictyosphaeria.

Hali

Percent of total annotated points annotated for Halimeda.

Lobph

Percent of total annotated points annotated for Lobophora.

Macro

Percent of total annotated points annotated for macroalgae.

Mdict

Percent of total annotated points annotated for Microdictyon.

Pad

Percent of total annotated points annotated for Padina.

Peysson

Percent of total annotated points annotated for Peyssonnelia.

Turf

Percent of total annotated points annotated for turf algae.

TURFH

Percent of total annotated points annotated for turf algae on hard substrate.

Unidentified

Percent of total annotated points annotated that were unidentified.